This PR moves all `httplib` code to a separate `HTTPUtil` class. This
class lives in the `DBConfig`, and has virtual methods. The idea is that
this class implements support for all HTTP requests we need. This class
can then be overridden by another class (i.e. `httpfs`), similar to how
the encryption util works. This has a number of advantages:

* Limiting HTTP requests to a single class means it is easier for WASM
to share code with the main library
* Limiting HTTP requests to a single class means it is easier to swap
HTTP libraries (e.g. if we want to switch to using curl)
* Having HTTP requests in a single class means it is possible for other
extensions besides httpfs to issue HTTP(S) requests without themselves
bundling httplib/openssl

For now this is only the extension installation code. The idea is to
build on top of this in the `HTTPFS` extension and to move all of the
web requests to the `HTTPUtil` class in a subsequent PR.

+++ src/include/duckdb/common/exception/http_exception.hpp
@@ -25,7 +25,8 @@ public:
 
 	template <class RESPONSE, typename ResponseShape<decltype(RESPONSE::status)>::status = 0, typename... ARGS>
 	explicit HTTPException(RESPONSE &response, const string &msg, ARGS... params)
-	    : HTTPException(response.status, response.body, response.headers, response.reason, msg, params...) {
+	    : HTTPException(static_cast<int>(response.status), response.body, response.headers, response.reason, msg,
+	                    params...) {
 	}
 
 	template <typename>
@@ -35,7 +36,8 @@ public:
 
 	template <class RESPONSE, typename ResponseWrapperShape<decltype(RESPONSE::code)>::code = 0, typename... ARGS>
 	explicit HTTPException(RESPONSE &response, const string &msg, ARGS... params)
-	    : HTTPException(response.code, response.body, response.headers, response.error, msg, params...) {
+	    : HTTPException(static_cast<int>(response.code), response.body, response.headers, response.error, msg,
+	                    params...) {
 	}
 
 	template <class HEADERS, typename... ARGS>
